Refactor content editors and endpoints
