+++ TEST_PLAN.md
@@ -1,0 +1,17 @@
+# Test Plan
+
+1. **Build and restore dependencies**
+   - Run `./setup.sh` to restore NuGet and client libraries.
+   - Ensure the command completes without errors.
+
+2. **Run unit tests**
+   - Execute `dotnet test website/MyWebApp.sln`.
+   - All tests should pass.
+
+3. **Manual UI smoke test**
+   - Start the application with `dotnet run --project website/MyWebApp`.
+   - Verify admin pages load correctly:
+     - Create and edit pages and sections using the unified section editor.
+     - Insert blocks via the updated API endpoints.
+   - Confirm media uploads still work and page layouts render as before.
+
