+++ website/MyWebApp/Controllers/AdminContentController.cs
@@ -7,8 +7,6 @@ using MyWebApp.Filters;
 using MyWebApp.Models;
 using MyWebApp.Services;
 using Microsoft.AspNetCore.Http;
-using Markdig;
-using System.IO;
 using System.Linq;
 
 namespace MyWebApp.Controllers;
@@ -18,13 +16,13 @@ public class AdminContentController : Controller
 {
     private readonly ApplicationDbContext _db;
     private readonly LayoutService _layout;
-    private readonly HtmlSanitizerService _sanitizer;
+    private readonly ContentProcessingService _content;
 
-    public AdminContentController(ApplicationDbContext db, LayoutService layout, HtmlSanitizerService sanitizer)
+    public AdminContentController(ApplicationDbContext db, LayoutService layout, ContentProcessingService content)
     {
         _db = db;
         _layout = layout;
-        _sanitizer = sanitizer;
+        _content = content;
     }
 
     private async Task LoadTemplatesAsync()
@@ -92,7 +90,7 @@ public class AdminContentController : Controller
                 s.Id = 0;
                 s.PageId = model.Id;
                 var file = files.FirstOrDefault(f => f.Name == $"Sections[{i}].File");
-                await PrepareHtmlAsync(s, file);
+                await _content.PrepareHtmlAsync(s, file);
                 _db.PageSections.Add(s);
             }
             await _db.SaveChangesAsync();
@@ -160,7 +158,7 @@ public class AdminContentController : Controller
                 s.Id = 0;
                 s.PageId = model.Id;
                 var file = files.FirstOrDefault(f => f.Name == $"Sections[{i}].File");
-                await PrepareHtmlAsync(s, file);
+                await _content.PrepareHtmlAsync(s, file);
                 _db.PageSections.Add(s);
             }
             await _db.SaveChangesAsync();
@@ -169,38 +167,6 @@ public class AdminContentController : Controller
         return RedirectToAction(nameof(Index));
     }
 
-    private async Task PrepareHtmlAsync(PageSection model, IFormFile? file)
-    {
-        switch (model.Type)
-        {
-            case PageSectionType.Html:
-                model.Html = _sanitizer.Sanitize(model.Html);
-                break;
-            case PageSectionType.Markdown:
-                var html = Markdig.Markdown.ToHtml(model.Html ?? string.Empty);
-                model.Html = _sanitizer.Sanitize(html);
-                break;
-            case PageSectionType.Code:
-                model.Html = $"<pre><code>{System.Net.WebUtility.HtmlEncode(model.Html)}</code></pre>";
-                break;
-            case PageSectionType.Image:
-            case PageSectionType.Video:
-                if (file != null && file.Length > 0)
-                {
-                    var uploads = Path.Combine("wwwroot", "uploads");
-                    Directory.CreateDirectory(uploads);
-                    var name = Path.GetFileName(file.FileName);
-                    var path = Path.Combine(uploads, name);
-                    using var stream = new FileStream(path, FileMode.Create);
-                    await file.CopyToAsync(stream);
-                    if (model.Type == PageSectionType.Image)
-                        model.Html = $"<img src='/uploads/{name}' alt='' />";
-                    else
-                        model.Html = $"<video controls src='/uploads/{name}'></video>";
-                }
-                break;
-        }
-    }
 
     public async Task<IActionResult> Delete(int id)
     {
