+++ website/MyWebApp/Controllers/AdminPageSectionController.cs
@@ -1,9 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Http;
-using Markdig;
-using System.IO;
 using MyWebApp.Data;
 using MyWebApp.Filters;
 using MyWebApp.Models;
@@ -16,13 +13,13 @@ public class AdminPageSectionController : Controller
 {
     private readonly ApplicationDbContext _db;
     private readonly LayoutService _layout;
-    private readonly HtmlSanitizerService _sanitizer;
+    private readonly ContentProcessingService _content;
 
-    public AdminPageSectionController(ApplicationDbContext db, LayoutService layout, HtmlSanitizerService sanitizer)
+    public AdminPageSectionController(ApplicationDbContext db, LayoutService layout, ContentProcessingService content)
     {
         _db = db;
         _layout = layout;
-        _sanitizer = sanitizer;
+        _content = content;
     }
 
     public async Task<IActionResult> Index(string? q)
@@ -64,7 +61,7 @@ public class AdminPageSectionController : Controller
             await LoadPagesAsync();
             return View(model);
         }
-        await PrepareHtmlAsync(model, file);
+        await _content.PrepareHtmlAsync(model, file);
         _db.PageSections.Add(model);
         await _db.SaveChangesAsync();
         _layout.Reset();
@@ -93,45 +90,13 @@ public class AdminPageSectionController : Controller
             await LoadPagesAsync();
             return View(model);
         }
-        await PrepareHtmlAsync(model, file);
+        await _content.PrepareHtmlAsync(model, file);
         _db.Update(model);
         await _db.SaveChangesAsync();
         _layout.Reset();
         return RedirectToAction(nameof(Index));
     }
 
-    private async Task PrepareHtmlAsync(PageSection model, IFormFile? file)
-    {
-        switch (model.Type)
-        {
-            case PageSectionType.Html:
-                model.Html = _sanitizer.Sanitize(model.Html);
-                break;
-            case PageSectionType.Markdown:
-                var html = Markdig.Markdown.ToHtml(model.Html ?? string.Empty);
-                model.Html = _sanitizer.Sanitize(html);
-                break;
-            case PageSectionType.Code:
-                model.Html = $"<pre><code>{System.Net.WebUtility.HtmlEncode(model.Html)}</code></pre>";
-                break;
-            case PageSectionType.Image:
-            case PageSectionType.Video:
-                if (file != null && file.Length > 0)
-                {
-                    var uploads = Path.Combine("wwwroot", "uploads");
-                    Directory.CreateDirectory(uploads);
-                    var name = Path.GetFileName(file.FileName);
-                    var path = Path.Combine(uploads, name);
-                    using var stream = new FileStream(path, FileMode.Create);
-                    await file.CopyToAsync(stream);
-                    if (model.Type == PageSectionType.Image)
-                        model.Html = $"<img src='/uploads/{name}' alt='' />";
-                    else
-                        model.Html = $"<video controls src='/uploads/{name}'></video>";
-                }
-                break;
-        }
-    }
 
     public async Task<IActionResult> Delete(int id)
     {
@@ -154,11 +119,4 @@ public class AdminPageSectionController : Controller
         return RedirectToAction(nameof(Index));
     }
 
-    [HttpGet]
-    public async Task<IActionResult> GetZonesForPage(int id)
-    {
-        var layout = await _db.Pages.Where(p => p.Id == id).Select(p => p.Layout).FirstOrDefaultAsync() ?? "single-column";
-        var zones = LayoutService.GetZones(layout);
-        return Json(zones);
-    }
 }
