+++ website/MyWebApp/Data/ApplicationDbContext.cs
@@ -111,8 +111,7 @@ namespace MyWebApp.Data
                     Type = PageSectionType.Html,
 
                     Html = "<div class=\"container-fluid nav-container\"><a class=\"logo\" href=\"/\">Screen Area Recorder Pro</a><nav class=\"site-nav\"><a href=\"/\">Home</a> {{nav}} <a href=\"/Download\">Download</a> <a href=\"/Home/Faq\">FAQ</a> <a href=\"/Home/Privacy\">Privacy</a> <a href=\"/Setup\">Setup</a> <a href=\"/Account/Login\">Login</a></nav></div>"
-                    ,
-                    ViewCount = 0
+                    
                 },
                 new PageSection
                 {
